requirements.txt
streamlit
pandas
numpy
requests
gspread
oauth2client
smartapi-python
pyotp
python-dotenv
logzero
websocket-client
